+++ src/Farmer/Arm/ContainerService.fs
@@ -67,7 +67,7 @@ type ManagedCluster =
                        {| agentPoolProfiles =
                            this.AgentPoolProfiles
                            |> List.mapi (fun idx agent ->
-                               {| name = if agent.Name = ResourceName.Empty then $"{this.Name.Value}-agent-pool{idx}"
+                               {| name = if agent.Name = ResourceName.Empty then $"nodepool{idx + 1}"
                                          else agent.Name.Value.ToLowerInvariant ()
                                   count = agent.Count
                                   maxPods = agent.MaxPods |> Option.toNullable